ToolScene UI추가 밑 UI 클릭 예외처리
TOOL_SCENEUI 클릭 예외처리 각 CUI의 자식UI들 매개변수로 부모에있는 중복된 mpTexture들고있어 삭제함 다이나믹 캐스트 삭제하고 강제형변환으로 변환 밑 각종 경고문구 제거

+++ Client/Random.cpp
@@ -14,7 +14,7 @@ CRandom::~CRandom()
 }
 void CRandom::Init()
 {
-	srand(GetTickCount64());
+	srand(static_cast<unsigned int>(GetTickCount64()));
 }
 
 int CRandom::Next(int start, int end)
