+++ Client/TextUI.cpp
@@ -17,7 +17,7 @@ TextUI::TextUI()
 	mStyle.fontSize = 32;
 	mStyle.color = RGB(0, 0, 0);
 	mStyle.boder = false;
-	mFont = CResMgr::GetI()->LoadFont(L"DungGeunMo", L"font\\DungGeunMo.ttf", mStyle.fontSize);
+	mFont = CResMgr::GetI()->LoadFont(L"DungGeunMo", L"font\\DungGeunMo.ttf", (int)mStyle.fontSize);
 }
 
 TextUI::~TextUI()
@@ -38,7 +38,7 @@ void TextUI::SetFontSize(float size)
 	mStyle.fontSize = size;
 	
 	// ReLoad Font
-	mFont = CResMgr::GetI()->LoadFont(L"DungGeunMo", L"font\\DungGeunMo.ttf", mStyle.fontSize);
+	mFont = CResMgr::GetI()->LoadFont(L"DungGeunMo", L"font\\DungGeunMo.ttf",(int)mStyle.fontSize);
 }
